refactor: 패키지, 클래스 명 적절하게 수정, 불필요한 상수 삭제
